Move internal/pubsub to correct location
