+++ internal/pubsub/message.go
