--- pubsub/internal/pubsub/README.md
@@ -1,8 +1,0 @@
-* Note: temp location only
-
-This folder will actually be located at `google-cloud-go/internal/pubsub`. It is
-here only for demonstration/prototyping purposes, to ensure pubsub tests pass.
-
-The files in this folder will be reviewed in a separate PR and incorporated in a
-new release of the `cloud.google.com/go` module, before updating the `pubsub`
-module.
